server_gql_django/requirements.txt
aniso8601==5.1.0
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
Django==2.2.2
django-cors-headers==3.0.2
django-graphiql==0.4.4
django-graphql-jwt==0.2.1
elasticsearch==6.4.0
graphene==2.1.5
graphene-django==2.3.0
graphql-core==2.2
graphql-relay==0.4.5
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
relativetimebuilder==2.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
